Add global nbgrader settings

+++ jupyterhub_config.py
@@ -59,11 +59,13 @@ c.JupyterHub.admin_access = True
 # SPAWNER
 # ---------------
 
-c.JupyterHub.spawner_class = 'systemdspawner.SystemdSpawner'
+c.JupyterHub.spawner_class = 'simplespawner.SimpleLocalProcessSpawner'
 
-c.SystemdSpawner.dynamic_users = True
+c.SimpleLocalProcessSpawner.home_path_template = '/home/{username}'
 
-c.SystemdSpawner.readwrite_paths = ['/srv/nbgrader/exchange']
+# c.SystemdSpawner.dynamic_users = True
+
+# c.SystemdSpawner.readwrite_paths = ['/srv/nbgrader/exchange']
 
 
 # c.Spawner.notebook_dir = '~/notebooks'
@@ -83,14 +85,21 @@ def pre_spawn_hook(spawner):
 #    except KeyError:
 #        subprocess.check_call(['useradd', '-ms', '/bin/bash', username])
 
-    with open(f'/home/{username}/nbgrader_config.py', 'w') as f:
-        f.write('''
+    if not os.path.exists(f'/home/{username}'):
+        os.makedirs(f'/home/{username}')
+
+    if spawner.userdata['role'] == 'Instructor':
+       if not os.path.exists(f'/home/{username}/nbgrader_config.py'):
+           spawner.log.info('Initialize instructor\'s directory. Create nbgrader_config and test course')
+           with open(f'/home/{username}/nbgrader_config.py', 'w') as f:
+               f.write(f'''
 c = get_config()
-c.CourseDirectory.course_id = "test_course"
-c.Exchange.root = "/srv/nbgrader/exchange"
+
+c.CourseDirectory.root = "/home/{username}/test_course"
+
+c.Exchange.path_includes_course = True
 ''')
-#     if spawner.userdata['role'] != 'Instructor':
-       # subprocess.check_call(f'jupyter nbextension disable --user formgrader/main --section=tree')
+           os.system(f'cd /home/{username} && nbgrader quickstart test_course')
        # subprocess.check_call(f'su - {username} && jupyter serverextension disable --user nbgrader.server_extensions.formgrader'.split())
 
 
@@ -103,6 +112,8 @@ def bind_auth_state(spawner, auth_state: dict) -> None:
 
 
 
-# c.Spawner.auth_state_hook = bind_auth_state
+c.Spawner.auth_state_hook = bind_auth_state
 
-# c.Spawner.pre_spawn_hook = pre_spawn_hook
+c.Spawner.pre_spawn_hook = pre_spawn_hook
+
+c.Spawner.args = ['--allow-root']
